requirements.txt
# Core Deep Learning
tensorflow>=2.10.0
keras>=2.10.0
tensorflow-hub>=0.12.0

# Web Framework
streamlit>=1.20.0

# Image Processing
Pillow>=9.0.0
opencv-python>=4.6.0

# Data Processing
numpy>=1.23.0
pandas>=1.5.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Machine Learning
scikit-learn>=1.2.0

# Jupyter (for notebook)
jupyter>=1.0.0
notebook>=6.5.0

# Utilities
python-dotenv>=0.21.0
